Move sisu export under final grade
