+++ client/src/components/course/course-results-view/SisuDownloadDialog.tsx
@@ -80,7 +80,7 @@ const SisuDownloadDialog = ({
     useState<DownloadOption>('unexported');
 
   const selectedStudents = selectedRows.map(row => ({
-    studentNumber: row.user.studentNumber!,
+    studentNumber: row.user.studentNumber,
     grades: row.finalGrades,
   }));
