+++ common/types/taskGrade.ts
@@ -6,15 +6,14 @@ import {z} from 'zod';
 
 import {AplusGradeSourceDataSchema} from './aplus';
 import {FinalGradeDataArraySchema} from './finalGrade';
-import {DateSchema, IdSchema, LanguageSchema} from './general';
-import {UserDataSchema} from './user';
+import {DateSchema, IdSchema} from './general';
+import {StudentDataSchema, TeacherDataSchema} from './user';
 
 const BaseGradeDataSchema = z.strictObject({
   gradeId: IdSchema,
-  grader: UserDataSchema,
+  grader: TeacherDataSchema,
   aplusGradeSource: AplusGradeSourceDataSchema.nullable(),
   grade: z.number(),
-  exportedToSisu: DateSchema.nullable(),
   date: DateSchema,
   expiryDate: DateSchema.nullable(),
   comment: z.string().nullable(),
@@ -55,7 +54,7 @@ export const CourseTaskGradesDataSchema = z.strictObject({
   grades: z.array(TaskGradeDataSchema),
 });
 export const StudentRowSchema = z.strictObject({
-  user: UserDataSchema,
+  user: StudentDataSchema,
   courseTasks: z.array(CourseTaskGradesDataSchema),
   finalGrades: FinalGradeDataArraySchema,
 });
@@ -66,11 +65,6 @@ export const LatestGradesSchema = z.array(
     date: DateSchema.nullable(),
   })
 );
-export const SisuCsvUploadSchema = z.strictObject({
-  assessmentDate: DateSchema.nullable(), // Assessment date override
-  completionLanguage: LanguageSchema.nullable(), // Defaults to course language
-  studentNumbers: z.array(z.string()).nonempty(),
-});
 
 export type TaskGradeData = z.infer<typeof TaskGradeDataSchema>;
 export type EditTaskGradeData = z.infer<typeof EditTaskGradeDataSchema>;
@@ -78,4 +72,3 @@ export type NewTaskGrade = z.infer<typeof NewTaskGradeSchema>;
 export type CourseTaskGradesData = z.infer<typeof CourseTaskGradesDataSchema>;
 export type StudentRow = z.infer<typeof StudentRowSchema>;
 export type LatestGrades = z.infer<typeof LatestGradesSchema>;
-export type SisuCsvUpload = z.infer<typeof SisuCsvUploadSchema>;
