+++ server/test/controllers/finalGrade.test.ts
@@ -2,6 +2,7 @@
 //
 // SPDX-License-Identifier: MIT
 
+import assert from 'assert';
 import supertest from 'supertest';
 
 import {
@@ -28,16 +29,24 @@ let editCourseId = -1;
 let editFinalGradeId = -1;
 let editCourseModelId = -1;
 let passFailCourseId = -1;
+let passFailCourseModelId = -1;
 let secondaryLangCourseId = -1;
+let secondaryLangCourseModelId = -1;
 let noRoleCourseId = -1;
 let noRoleFinalGradeId = -1;
 const nonExistentId = 1000000;
+
+const nonExistentStudentNumber = '945942';
+const students: {id: number; studentNumber: string; finalGrade: number}[] = [];
+let studentNumbers: string[] = [];
 
 beforeAll(async () => {
   cookies = await getCookies();
 
   let gradingModelId;
-  [courseId, , , gradingModelId] = await createData.createCourse({});
+  [courseId, , , gradingModelId] = await createData.createCourse({
+    courseData: {maxCredits: 5, courseCode: 'CS-84389'},
+  });
   // Create 10 final grades
   for (let i = 0; i < 10; i++) {
     const student = await createData.createUser();
@@ -67,12 +76,23 @@ beforeAll(async () => {
     TEACHER_ID
   );
 
-  [passFailCourseId] = await createData.createCourse({
-    courseData: {gradingScale: GradingScale.PassFail},
-  });
-  [secondaryLangCourseId] = await createData.createCourse({
-    courseData: {gradingScale: GradingScale.SecondNationalLanguage},
-  });
+  [passFailCourseId, , , passFailCourseModelId] = await createData.createCourse(
+    {
+      courseData: {
+        gradingScale: GradingScale.PassFail,
+        maxCredits: 5,
+        courseCode: 'CS-A9542',
+      },
+    }
+  );
+  [secondaryLangCourseId, , , secondaryLangCourseModelId] =
+    await createData.createCourse({
+      courseData: {
+        gradingScale: GradingScale.SecondNationalLanguage,
+        maxCredits: 5,
+        courseCode: 'CS-A8341',
+      },
+    });
 
   [noRoleCourseId] = await createData.createCourse({
     hasTeacher: false,
@@ -85,6 +105,52 @@ beforeAll(async () => {
     null,
     TEACHER_ID
   );
+
+  // Create students & final grades for Sisu CSV test
+  for (let i = 0; i < 10; i++) {
+    const newUser = await createData.createUser();
+    assert(newUser.studentNumber); // Fix typescript warning
+    students.push({
+      id: newUser.id,
+      studentNumber: newUser.studentNumber,
+      finalGrade: Math.floor(Math.random() * 6),
+    });
+  }
+  studentNumbers = students.map(student => student.studentNumber);
+
+  for (const student of students) {
+    // Create a worse final grade before the actual one
+    if (student.finalGrade > 0) {
+      await createData.createFinalGrade(
+        courseId,
+        student.id,
+        gradingModelId,
+        TEACHER_ID,
+        student.finalGrade -
+          Math.floor(Math.random() * (student.finalGrade + 1))
+      );
+    }
+
+    await createData.createFinalGrade(
+      courseId,
+      student.id,
+      gradingModelId,
+      TEACHER_ID,
+      student.finalGrade
+    );
+
+    // Create a worse final grade after the actual one
+    if (student.finalGrade > 0) {
+      await createData.createFinalGrade(
+        courseId,
+        student.id,
+        gradingModelId,
+        TEACHER_ID,
+        student.finalGrade -
+          Math.floor(Math.random() * (student.finalGrade + 1))
+      );
+    }
+  }
 });
 
 afterAll(async () => {
@@ -496,5 +562,232 @@ describe('Test Delete/v1/courses/:courseId/final-grades/:finalGradeId - delete a
   it('should respond with 409 when grade does not belong to course', async () => {
     const url = `/v1/courses/${editCourseId}/final-grades/${noRoleFinalGradeId}`;
     await responseTests.testConflict(url, cookies.adminCookie).delete();
+  });
+});
+
+describe('Test POST /v1/courses/:courseId/grades/csv/sisu - export Sisu compatible grading in CSV', () => {
+  const today = new Date().toLocaleDateString('FI-fi');
+  const createCSVString = (
+    studentData: {studentNumber: string; finalGrade: number}[],
+    dateStr: string,
+    language: string
+  ): string[] => {
+    const data = [];
+    for (const student of studentData) {
+      const csvStudent = `${student.studentNumber},${student.finalGrade}`;
+      data.push(`${csvStudent},5,${dateStr},${language}`);
+    }
+    return data;
+  };
+
+  it('should export CSV', async () => {
+    const testCookies = [cookies.adminCookie, cookies.teacherCookie];
+    for (const cookie of testCookies) {
+      const res = await request
+        .post(`/v1/courses/${courseId}/grades/csv/sisu`)
+        .send({studentNumbers, assessmentDate: null, completionLanguage: null})
+        .set('Cookie', cookie)
+        .set('Accept', 'text/csv')
+        .expect(HttpCode.Ok);
+
+      expect(res.text)
+        .toBe(`studentNumber,grade,credits,assessmentDate,completionLanguage,comment
+${createCSVString(students, today, 'en').join(',\n')},\n`);
+      expect(res.headers['content-disposition']).toBe(
+        'attachment; filename="final_grades_course_CS-84389_' +
+          `${new Date().toLocaleDateString('fi-FI')}.csv"`
+      );
+    }
+  });
+
+  it('should export only selected grades', async () => {
+    const selectedStudents = [students[0], students[3], students[6]];
+    const res = await request
+      .post(`/v1/courses/${courseId}/grades/csv/sisu`)
+      .send({
+        studentNumbers: selectedStudents.map(student => student.studentNumber),
+        assessmentDate: null,
+        completionLanguage: null,
+      })
+      .set('Cookie', cookies.adminCookie)
+      .set('Accept', 'text/csv')
+      .expect(HttpCode.Ok);
+
+    expect(res.text)
+      .toBe(`studentNumber,grade,credits,assessmentDate,completionLanguage,comment
+${createCSVString(selectedStudents, today, 'en').join(',\n')},\n`);
+    expect(res.headers['content-disposition']).toBe(
+      'attachment; filename="final_grades_course_CS-84389_' +
+        `${new Date().toLocaleDateString('fi-FI')}.csv"`
+    );
+  });
+
+  it('should export CSV successfully with custom assessmentDate and completionLanguage', async () => {
+    const res = await request
+      .post(`/v1/courses/${courseId}/grades/csv/sisu`)
+      .send({
+        studentNumbers,
+        assessmentDate: new Date(2023, 4, 12),
+        completionLanguage: 'JA',
+      })
+      .set('Cookie', cookies.adminCookie)
+      .set('Accept', 'text/csv')
+      .expect(HttpCode.Ok);
+
+    expect(res.text)
+      .toBe(`studentNumber,grade,credits,assessmentDate,completionLanguage,comment
+${createCSVString(students, '12.5.2023', 'ja').join(',\n')},\n`);
+    expect(res.headers['content-disposition']).toBe(
+      'attachment; filename="final_grades_course_CS-84389_' +
+        `${new Date().toLocaleDateString('fi-FI')}.csv"`
+    );
+  });
+
+  it('should export CSV when gradingScale is pass/fail', async () => {
+    const studentNums = [];
+    for (let i = 0; i <= 1; i++) {
+      studentNums.push(students[i].studentNumber);
+      await createData.createFinalGrade(
+        passFailCourseId,
+        students[i].id,
+        passFailCourseModelId,
+        TEACHER_ID,
+        i
+      );
+    }
+    const res = await request
+      .post(`/v1/courses/${passFailCourseId}/grades/csv/sisu`)
+      .send({
+        studentNumbers: studentNums,
+        assessmentDate: null,
+        completionLanguage: null,
+      })
+      .set('Cookie', cookies.teacherCookie)
+      .set('Accept', 'text/csv')
+      .expect(HttpCode.Ok);
+
+    expect(res.text)
+      .toBe(`studentNumber,grade,credits,assessmentDate,completionLanguage,comment
+${students[0].studentNumber},fail,5,${today},en,
+${students[1].studentNumber},pass,5,${today},en,\n`);
+    expect(res.headers['content-disposition']).toBe(
+      'attachment; filename="final_grades_course_CS-A9542_' +
+        `${new Date().toLocaleDateString('fi-FI')}.csv"`
+    );
+  });
+
+  it('should export CSV when gradingScale is secondary language', async () => {
+    const studentNums = [];
+    for (let i = 0; i <= 2; i++) {
+      studentNums.push(students[i].studentNumber);
+      await createData.createFinalGrade(
+        secondaryLangCourseId,
+        students[i].id,
+        secondaryLangCourseModelId,
+        TEACHER_ID,
+        i
+      );
+    }
+    const res = await request
+      .post(`/v1/courses/${secondaryLangCourseId}/grades/csv/sisu`)
+      .send({
+        studentNumbers: studentNums,
+        assessmentDate: null,
+        completionLanguage: null,
+      })
+      .set('Cookie', cookies.teacherCookie)
+      .set('Accept', 'text/csv')
+      .expect(HttpCode.Ok);
+
+    expect(res.text)
+      .toBe(`studentNumber,grade,credits,assessmentDate,completionLanguage,comment
+${students[0].studentNumber},Fail,5,${today},en,
+${students[1].studentNumber},SAT,5,${today},en,
+${students[2].studentNumber},G,5,${today},en,\n`);
+    expect(res.headers['content-disposition']).toBe(
+      'attachment; filename="final_grades_course_CS-A8341_' +
+        `${new Date().toLocaleDateString('fi-FI')}.csv"`
+    );
+  });
+
+  it('should respond with 400 if validation fails', async () => {
+    const url = `/v1/courses/${courseId}/grades/csv/sisu`;
+    const badRequest = responseTests.testBadRequest(url, cookies.adminCookie);
+
+    await badRequest.post({
+      studentNumbers,
+      assessmentDate: null,
+      completionLanguage: 'ja',
+    });
+    await badRequest.post({
+      studentNumbers,
+      assessmentDate: '2024-12-00T00:00:00.000Z',
+    });
+  });
+
+  it('should respond with 400 if id is invalid', async () => {
+    const url = '/v1/courses/bad/grades/csv/sisu';
+    const data = {
+      studentNumbers,
+      assessmentDate: null,
+      completionLanguage: null,
+    };
+    await responseTests.testBadRequest(url, cookies.adminCookie).post(data);
+  });
+
+  it('should respond with 401 or 403 if not authorized', async () => {
+    let url = `/v1/courses/${courseId}/grades/csv/sisu`;
+    const data = {
+      studentNumbers,
+      assessmentDate: null,
+      completionLanguage: null,
+    };
+    await responseTests.testUnauthorized(url).post(data);
+
+    await responseTests
+      .testForbidden(url, [cookies.assistantCookie, cookies.studentCookie])
+      .post(data);
+
+    url = `/v1/courses/${noRoleCourseId}/grading-models`;
+    await responseTests
+      .testForbidden(url, [
+        cookies.teacherCookie,
+        cookies.assistantCookie,
+        cookies.studentCookie,
+      ])
+      .post(data);
+  });
+
+  it('should respond with 404 if grades have not been calculated yet', async () => {
+    const url = `/v1/courses/${noRoleCourseId}/grades/csv/sisu`;
+    const data = {
+      studentNumbers: [studentNumbers[0]],
+      assessmentDate: null,
+      completionLanguage: null,
+    };
+
+    await responseTests.testNotFound(url, cookies.adminCookie).post(data);
+  });
+
+  it('should respond with 404 if student number not found', async () => {
+    const url = `/v1/courses/${noRoleCourseId}/grades/csv/sisu`;
+    const data = {
+      studentNumbers: [nonExistentStudentNumber],
+      assessmentDate: null,
+      completionLanguage: null,
+    };
+
+    await responseTests.testNotFound(url, cookies.adminCookie).post(data);
+  });
+
+  it('should respond with 404 if not found', async () => {
+    const url = `/v1/courses/${nonExistentId}/grades/csv/sisu`;
+    const data = {
+      studentNumbers,
+      assessmentDate: null,
+      completionLanguage: null,
+    };
+
+    await responseTests.testNotFound(url, cookies.adminCookie).post(data);
   });
 });
